Initial commit of CentOS stemcell-builder support.  * Based on work from https://github.com/pmenglund/bosh/tree/centos  * Added spec centos_stemcell-vsphere.spec which utilizes new centos stages and existing stages  * Initializes and builds the chroot root fs using rpm / yum  * Added a variety of stages named centos_* which implement centos specific implementations of original ubuntu stages  * CentOS vsphere stem cell can be built by specifying the "CENTOS_ISO" and "base_os=centos" vars in settings.bash.

+++ stemcell_builder/lib/helpers.sh
@@ -1,29 +1,35 @@
 # Copyright (c) 2009-2012 VMware, Inc.
 
+base_os="centos"
+
 function codename() {
   if [ -r /etc/lsb-release ]; then
-    source /etc/lsb-release
-    if [ -n "${DISTRIB_CODENAME}" ]; then
-      echo ${DISTRIB_CODENAME}
-      return 0
-    fi
+	source /etc/lsb-release
+	if [ -n "${DISTRIB_CODENAME}" ]; then
+	  echo ${DISTRIB_CODENAME}
+	  return 0
+	fi
   else
-    lsb_release -cs
+	lsb_release -cs
   fi
 }
 
 function disable {
-  mv $1 $1.back
-  ln -s /bin/true $1
+  if [ ! "${base_os}" = "centos" ] ; then
+	  mv $1 $1.back
+	  ln -s /bin/true $1
+  fi
 }
 
 function enable {
-  if [ -L $1 ]
-  then
-    mv $1.back $1
-  else
-    # No longer a symbolic link, must have been overwritten
-    rm -f $1.back
+  if [ ! "${base_os}" = "centos" ] ; then
+	  if [ -L $1 ]
+	  then
+		  mv $1.back $1
+	  else
+		  # No longer a symbolic link, must have been overwritten
+		  rm -f $1.back
+	  fi
   fi
 }
 
@@ -36,14 +42,14 @@ function run_in_chroot {
   disable $chroot/usr/sbin/invoke-rc.d
 
   unshare -m $SHELL <<EOS
-    mkdir -p $chroot/dev
-    mount -n --bind /dev $chroot/dev
-    mount -n --bind /dev/pts $chroot/dev/pts
+	mkdir -p $chroot/dev
+	mount -n --bind /dev $chroot/dev
+	mount -n --bind /dev/pts $chroot/dev/pts
 
-    mkdir -p $chroot/proc
-    mount -n -t proc proc $chroot/proc
+	mkdir -p $chroot/proc
+	mount -n -t proc proc $chroot/proc
 
-    chroot $chroot env -i $(cat $chroot/etc/environment) http_proxy=${http_proxy:-} bash -e -c "$script"
+	chroot $chroot env -i $(cat $chroot/etc/environment) http_proxy=${http_proxy:-} bash -e -c "$script"
 EOS
 
   # Enable daemon startup
@@ -56,7 +62,7 @@ declare -a on_exit_items
 function on_exit {
   for i in "${on_exit_items[@]}"
   do
-    eval $i
+	eval $i
   done
 }
 
@@ -64,6 +70,6 @@ function add_on_exit {
   local n=${#on_exit_items[*]}
   on_exit_items[$n]="$*"
   if [[ $n -eq 0 ]]; then
-    trap on_exit EXIT
+	trap on_exit EXIT
   fi
 }
